Dual-port Memory
We have two memory chips which are controlled by CS1 and CS2.
For instance, if CS1 = ‘1’, this means that the first port is active. Now if WE1 = ‘1’, the data,
Data1 is going to be written in the cell which Addr1 is referring to.
If WE1 = ‘0’, the data on the memory cell which Addr1 is referring to will be written on
Data1.
The same rules hold for the second chip.