Elevator System
An elevator system using a finite state machine to go from one floor to another and wait there for a specified time. The plan for the elevator is written into the memory and interrupt occurs when the button on one floor is pushed while the elevator is doing its routine. For creating the interrupt, a random number generator circuit is used.